[spv-in] move all the image-related logic into a sub-module
